pyproject.toml
[project]
name = "genai-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.100.0",
    "fitz>=0.0.1.dev2",
    "frontend>=0.0.3",
    "gradio>=4.0.0",
    "httpx>=0.24.0",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-llms-openai>=0.6.8",
    "llama-index-readers-file>=0.5.4",
    "llama-index-vector-stores-duckdb>=0.5.1",
    "loguru>=0.7.3",
    "uvicorn[standard]>=0.24.0",
]
